requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioresponses==0.7.8
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
coverage==7.9.1
fastapi==0.115.13
freezegun==1.5.2
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
multidict==6.5.0
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
respx==0.22.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
yarl==1.20.1
